app/requirements.txt
Django==3.2
django-admin-interface==0.15.5
gunicorn==20.0.4
Pillow==8.1.2
psycopg2-binary==2.8.6
django-taggit==1.3.0
